Array lenght must be uint

+++ src/lib.rs
@@ -15,14 +15,14 @@ use syntax::parse::token;
 
 #[deriving(Show)]
 enum Field {
-   ArrayField {name: String, count: u64, element_length: u64},
+   ArrayField {name: String, count: uint, element_length: u64},
    ScalarField {name: String, length: u64},   
 }
 
 impl Field {
     fn bit_len(&self) -> u64 {
        match *self {
-          ArrayField(_, count, element_length) => count * element_length,
+          ArrayField(_, count, element_length) => (count as u64) * element_length,
           ScalarField(_, length) => length,
        }
     }
@@ -55,7 +55,7 @@ fn parse_field(parser: &mut Parser) -> Field {
           parser.span_fatal(span, "Field length must be > 0");
        }
        parser.expect(&token::CloseDelim(token::Bracket));
-       ArrayField {name: name,  element_length:  element_length, count: count}
+       ArrayField {name: name,  element_length:  element_length, count: count as uint}
     }
     else {
       //ScalarField
@@ -82,7 +82,7 @@ fn expand_bitfield(cx: &mut ExtCtxt, _sp: Span, tts: &[ast::TokenTree])
     let fields = parser.parse_seq_to_end(&token::Eof, sep, |p| parse_field(p));
 
     let bit_length = fields.iter().fold(0, |a, b| a + b.bit_len());
-    let byte_length = (bit_length+7)/8;
+    let byte_length = ((bit_length+7)/8) as uint;
     
     let struct_decl = quote_item!(cx, struct $struct_ident<'a> { data: &'a [u8, ..$byte_length]};).unwrap();
